Refactory service user
