+++ src/CoffeeShopOrder.Application/Interfaces/IUserService.cs
@@ -1,11 +1,12 @@
-﻿using CoffeeShopOrder.Application.Models.Requests;
+﻿using CoffeeShopOrder.Application.Models.DTO;
+using CoffeeShopOrder.Application.Models.Requests;
 using CoffeeShopOrder.Application.Models.Responses;
 
 namespace CoffeeShopOrder.Application.Interfaces
 {
     public interface IUserService
     {
-        Task<CreateUserResponse> Create(UserCreateRequest userCreateRequest, CancellationToken cancellationToken);
+        Task<UserDTO> Create(UserCreateRequest userCreateRequest, CancellationToken cancellationToken);
         Task<GetUserResponse> Get(Guid id, CancellationToken cancellationToken);
     }
 }
