--- src/CoffeeShopOrder.Application/Models/Responses/CreateUserResponse.cs
@@ -1,9 +1,0 @@
-﻿using CoffeeShopOrder.Application.Models.DTO;
-
-namespace CoffeeShopOrder.Application.Models.Responses
-{
-    public record CreateUserResponse
-    {
-        UserDTO User { get; set; }
-    }
-}
