+++ src/CoffeeShopOrder.Application/Services/UserService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using CoffeeShopOrder.Application.Interfaces;
+using CoffeeShopOrder.Application.Models.DTO;
 using CoffeeShopOrder.Application.Models.Requests;
 using CoffeeShopOrder.Application.Models.Responses;
 using CoffeeShopOrder.Domain.Entities;
@@ -20,14 +21,14 @@ namespace CoffeeShopOrder.Application.Services
             _mapper = mapper;
         }
 
-        public async Task<CreateUserResponse> Create(UserCreateRequest userCreateRequest, CancellationToken cancellationToken)
+        public async Task<UserDTO> Create(UserCreateRequest userCreateRequest, CancellationToken cancellationToken)
         {
             User user = _mapper.Map<User>(userCreateRequest);
 
             _userRepository.Create(user);
             await _unitOfWork.Commit(cancellationToken);
 
-            return _mapper.Map<CreateUserResponse>(user);
+            return _mapper.Map<UserDTO>(user);
         }
 
         public async Task<GetUserResponse> Get(Guid id, CancellationToken cancellationToken)
