+++ src/CoffeeShopOrder.Domain/Common/EntityBase.cs
@@ -11,6 +11,7 @@
         {
             Id = Guid.NewGuid();
             CreatedAt = DateTime.UtcNow;
+            UpdatedAt = DateTime.UtcNow;
         }
 
         public void SetUpdated()
