+++ src/CoffeeShopOrder.Infrastructure/Context/AppDbContext.cs
@@ -7,8 +7,8 @@ namespace CoffeeShopOrder.Application.Context
     {
         public AppDbContext(DbContextOptions<AppDbContext> options) : base(options) {}
 
-        public DbSet<User> Users { get; set; }
-        public DbSet<Order> Orders { get; set; }
-        public DbSet<Coffee> Coffees { get; set; }
+        public DbSet<User> User { get; set; }
+        public DbSet<Order> Order { get; set; }
+        public DbSet<Coffee> Coffee { get; set; }
     }
 }
