+++ src/CoffeeShopOrder.Infrastructure/Extensions/ServiceExtensions.cs
@@ -6,14 +6,14 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace CoffeeShopOrder.Infrastructure
+namespace CoffeeShopOrder.Infrastructure.Extensions
 {
     public static class ServiceExtensions
     {
         public static void ConfigurePersistenceApp(this IServiceCollection services,
                                                     IConfiguration configuration)
         {
-            var connectionString = configuration.GetConnectionString("SqlServer");
+            var connectionString = configuration.GetConnectionString("SqlServerConnection");
             services.AddDbContext<AppDbContext>(opt => opt.UseSqlServer(connectionString));
 
             services.AddScoped<IUnitOfWork, UnitOfWork>();
